funding_checker/requirements.txt
requests
python-telegram-bot
pymongo
aiohttp
python-dotenv
